requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
arabic-reshaper==2.1.3
    # via xhtml2pdf
asgiref==3.5.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via ipython
certifi==2022.6.15.1
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via pyhanko
coverage==6.4.4
    # via -r requirements.in
crispy-bootstrap5==0.6
    # via -r requirements.in
cryptography==38.0.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.6.0
    # via svglib
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.1.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-admin-autocomplete-filter
    #   django-browser-reload
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-user-visit
    #   djangorestframework
django-admin-autocomplete-filter==0.7.1
    # via -r requirements.in
django-admin-rangefilter==0.8.8
    # via -r requirements.in
django-browser-reload==1.6.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via crispy-bootstrap5
django-debug-toolbar==3.6.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-simple-history==3.1.1
    # via -r requirements.in
django-user-visit==0.5.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==1.0.0
    # via stack-data
future==0.18.2
    # via arabic-reshaper
html5lib==1.1
    # via xhtml2pdf
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
ipython==8.5.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
lxml==4.9.1
    # via svglib
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.6
    # via ipython
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==21.3
    # via
    #   pytest
    #   pytest-sugar
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via
    #   -r requirements.in
    #   reportlab
    #   xhtml2pdf
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via ipython
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via ipython
pyhanko==0.13.2
    # via xhtml2pdf
pyhanko-certvalidator==0.19.5
    # via
    #   pyhanko
    #   xhtml2pdf
pyparsing==3.0.9
    # via packaging
pypdf3==1.0.6
    # via xhtml2pdf
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements.in
pytest-sugar==0.9.5
    # via -r requirements.in
python-bidi==0.4.2
    # via xhtml2pdf
pytz==2022.2.1
    # via
    #   djangorestframework
    #   pyhanko
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   pyhanko
    #   tablib
qrcode==7.3.1
    # via pyhanko
reportlab==3.6.11
    # via
    #   svglib
    #   xhtml2pdf
requests==2.28.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
    #   rollbar
rollbar==0.16.3
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   html5lib
    #   python-bidi
    #   rollbar
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.5.0
    # via ipython
svglib==1.4.1
    # via xhtml2pdf
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
termcolor==1.1.0
    # via pytest-sugar
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via pytest
tqdm==4.64.1
    # via pypdf3
traitlets==5.3.0
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2022.2
    # via pytz-deprecation-shim
tzlocal==4.2
    # via pyhanko
ua-parser==0.16.1
    # via user-agents
uritools==4.0.0
    # via pyhanko-certvalidator
urllib3==1.26.12
    # via requests
user-agents==2.2.0
    # via django-user-visit
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise==6.2.0
    # via -r requirements.in
xhtml2pdf==0.2.8
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
